lab2 finished

+++ lab2/DebugPractice/DebugExercise1.java
@@ -4,7 +4,9 @@
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+//        int quotient = top / bottom;
+        // txy-modify
+        float quotient = (float)top / bottom;
         int result = Math.round(quotient);
         return result;
     }
